load_tests/requirements-locust.txt
locust==2.40.2
Pillow==11.3.0
